[popoto_driver] Convertefd the goby header from 2 byte to 1 byte. Will discuss in the future when time permits what the final decision will be in terms of goby header size for popoto transmission and then formalise this change

+++ src/acomms/modemdriver/popoto_driver.h
@@ -72,7 +72,9 @@ class PopotoDriver : public ModemDriverBase
     void parse_in(const std::string& in, std::map<std::string, std::string>* out);
     void signal_and_write(const std::string& raw);
 
-    std::uint16_t CreateGobyHeader(const protobuf::ModemTransmission& m);
+    // 2 byte header code
+    // std::uint16_t CreateGobyHeader(const protobuf::ModemTransmission& m);
+    std::uint8_t CreateGobyHeader(const protobuf::ModemTransmission& m);
     void DecodeGobyHeader(std::uint8_t header, std::uint8_t ack_num,protobuf::ModemTransmission& m);
     void DecodeHeader(std::vector<uint8_t> data, protobuf::ModemTransmission& m);
     void ProcessJSON(const std::string& message,protobuf::ModemTransmission& modem_msg);
